Renamin domain normalizers and refacto the DI a bit
